add int in noise creation

+++ pypulse/PulseGeneration.py
@@ -274,8 +274,8 @@ def noise_pulse(sampling_rate, params):
     pulse = (np.array(signal.square(2 * np.pi * params['shatter_frequency'] * t, duty=guide_pulse)) / 2) + 0.5
 
     # Attach onset and offset
-    onset = np.zeros(sampling_rate * params['onset'])
-    offset = np.zeros(sampling_rate * params['offset'])
+    onset = np.zeros(int(sampling_rate * params['onset']))
+    offset = np.zeros(int(sampling_rate * params['offset']))
 
     total_length = round(duration + params['onset'] + params['offset'], 10)
     return np.hstack((onset, pulse, offset)), np.linspace(0, total_length, total_length * sampling_rate)
